+++ raphtory/src/algorithms/mod.rs
@@ -38,6 +38,4 @@ pub mod pagerank;
 pub mod reciprocity;
 pub mod triangle_count;
 pub mod triplet_count;
-pub mod three_node_local;
-pub mod three_node_motifs;
 pub mod motifs;
