+++ raphtory/src/algorithms/motifs/mod.rs
@@ -1,0 +1,2 @@
+pub mod three_node_local;
+pub mod three_node_motifs;
